requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
aiosmtplib==4.0.2
alembic==1.13.3
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
babel==2.17.0
bcrypt==4.2.0
black==24.10.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
cryptography==46.0.1
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
flake8==7.1.1
ghp-import==2.1.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.44
mkdocs-material-extensions==1.3.1
mypy==1.13.0
mypy_extensions==1.1.0
numpy==2.3.3
opencv-python==4.10.0.84
packaging==25.0
paginate==0.5.7
passlib==1.7.4
pathlib2==2.3.7
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.4.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16.1
PyMySQL==1.1.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.12
PyYAML==6.0.3
pyyaml_env_tag==1.1
regex==2025.9.18
requests==2.32.3
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.39.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
structlog==24.4.0
tqdm==4.66.6
typer==0.19.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.32.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
#python-multipart
#tensorflow
#mediapipe
#opencv-python-headless
#numpy
